Move pessimistic locking settings to dedicated class
